refactor(images): extract shared image validation from enrollment
